+++ UiLepra/Source/Win32/UiWin32DisplayManager.cpp
@@ -613,7 +613,7 @@ bool Win32DisplayManager::OnMessage(int pMsg, int pwParam, long plParam)
 	{
 		case WM_SIZING:
 		{
-			LPRECT lRect = (LPRECT)(uint64)plParam;
+			LPRECT lRect = (LPRECT)(intptr_t)plParam;
 
 			int lClientWidth  = GetClientWidth(lRect->right - lRect->left);
 			int lClientHeight = GetClientHeight(lRect->bottom - lRect->top);
