+++ UiLepra/Source/X11/UiX11DisplayManager.cpp
@@ -554,7 +554,7 @@ Bool X11DisplayManager::OnMessage(XEvent* e)
 	{
 		case WM_SIZING:
 		{
-			LPRECT lRect = (LPRECT)(uint64)plParam;
+			LPRECT lRect = (LPRECT)(intptr_t)plParam;
 
 			int lClientWidth  = GetClientWidth(lRect->right - lRect->left);
 			int lClientHeight = GetClientHeight(lRect->bottom - lRect->top);
